Add pytest plugin for filtering of integration tests
