+++ tests/integration/install/test_logs.py
