+++ tests/integration/workspace_access/test_workflows.py
@@ -1,45 +1,15 @@
-import logging
 from datetime import timedelta
 
-import pytest  # pylint: disable=wrong-import-order
+import pytest
 from databricks.labs.blueprint.parallel import ManyError
-from databricks.sdk.errors import (
-    InvalidParameterValue,
-    NotFound,
-)
+
+from databricks.sdk.errors import NotFound, InvalidParameterValue
 from databricks.sdk.retries import retried
 from databricks.sdk.service import sql
 from databricks.sdk.service.iam import PermissionLevel
 from databricks.sdk.service.workspace import AclPermission
 
-from tests.integration.conftest import modified_or_skip
 
-logger = logging.getLogger(__name__)
-
-
-@modified_or_skip("assessment")
-@retried(on=[NotFound, InvalidParameterValue], timeout=timedelta(minutes=8))
-def test_running_real_assessment_job(ws, installation_ctx, make_cluster_policy, make_cluster_policy_permissions):
-
-    ctx = installation_ctx.replace(skip_dashboards=False)
-    ws_group_a, _ = ctx.make_ucx_group()
-
-    cluster_policy = make_cluster_policy()
-    make_cluster_policy_permissions(
-        object_id=cluster_policy.policy_id,
-        permission_level=PermissionLevel.CAN_USE,
-        group_name=ws_group_a.display_name,
-    )
-    ctx.__dict__['include_object_permissions'] = [f"cluster-policies:{cluster_policy.policy_id}"]
-    ctx.workspace_installation.run()
-
-    ctx.deployed_workflows.run_workflow("assessment")
-
-    after = ctx.generic_permissions_support.load_as_dict("cluster-policies", cluster_policy.policy_id)
-    assert after[ws_group_a.display_name] == PermissionLevel.CAN_USE
-
-
-@modified_or_skip("workspace_access")
 @retried(on=[NotFound, InvalidParameterValue], timeout=timedelta(minutes=8))
 def test_running_real_migrate_groups_job(
     ws,
@@ -81,7 +51,6 @@ def test_running_real_migrate_groups_job(
     assert found[f"{installation_ctx.config.renamed_group_prefix}{ws_group_a.display_name}"] == PermissionLevel.CAN_USE
 
 
-@modified_or_skip("workspace_access")
 @retried(on=[NotFound, InvalidParameterValue], timeout=timedelta(minutes=5))
 def test_running_real_validate_groups_permissions_job(
     installation_ctx,
@@ -128,7 +97,6 @@ def test_running_real_validate_groups_permissions_job(
     installation_ctx.deployed_workflows.run_workflow("validate-groups-permissions")
 
 
-@modified_or_skip("workspace_access")
 @retried(on=[NotFound], timeout=timedelta(minutes=8))
 def test_running_real_validate_groups_permissions_job_fails(
     ws, installation_ctx, make_cluster_policy, make_cluster_policy_permissions
@@ -156,47 +124,3 @@ def test_running_real_validate_groups_permissions_job_fails(
 
     with pytest.raises(ManyError):
         installation_ctx.deployed_workflows.run_workflow("validate-groups-permissions")
-
-
-@modified_or_skip("hive_metastore")
-@retried(on=[NotFound], timeout=timedelta(minutes=8))
-@pytest.mark.parametrize('prepare_tables_for_migration', [('hiveserde')], indirect=True)
-def test_hiveserde_table_in_place_migration_job(ws, installation_ctx, prepare_tables_for_migration):
-    tables, dst_schema = prepare_tables_for_migration
-    ctx = installation_ctx.replace(
-        extend_prompts={
-            r".*Do you want to update the existing installation?.*": 'yes',
-        },
-    )
-    ctx.workspace_installation.run()
-    ctx.deployed_workflows.run_workflow("migrate-external-hiveserde-tables-in-place-experimental")
-    # assert the workflow is successful
-    assert ctx.deployed_workflows.validate_step("migrate-external-hiveserde-tables-in-place-experimental")
-    # assert the tables are migrated
-    for table in tables.values():
-        try:
-            assert ws.tables.get(f"{dst_schema.catalog_name}.{dst_schema.name}.{table.name}").name
-        except NotFound:
-            assert False, f"{table.name} not found in {dst_schema.catalog_name}.{dst_schema.name}"
-
-
-@modified_or_skip("hive_metastore")
-@retried(on=[NotFound], timeout=timedelta(minutes=8))
-@pytest.mark.parametrize('prepare_tables_for_migration', [('hiveserde')], indirect=True)
-def test_hiveserde_table_ctas_migration_job(ws, installation_ctx, prepare_tables_for_migration):
-    tables, dst_schema = prepare_tables_for_migration
-    ctx = installation_ctx.replace(
-        extend_prompts={
-            r".*Do you want to update the existing installation?.*": 'yes',
-        },
-    )
-    ctx.workspace_installation.run()
-    ctx.deployed_workflows.run_workflow("migrate-external-tables-ctas")
-    # assert the workflow is successful
-    assert ctx.deployed_workflows.validate_step("migrate-external-tables-ctas")
-    # assert the tables are migrated
-    for table in tables.values():
-        try:
-            assert ws.tables.get(f"{dst_schema.catalog_name}.{dst_schema.name}.{table.name}").name
-        except NotFound:
-            assert False, f"{table.name} not found in {dst_schema.catalog_name}.{dst_schema.name}"
